adding custom config templates with detailed params

+++ engine/call.py
@@ -65,6 +65,7 @@ if args.refine_config is not None:
     refine_config = config_utils.load_config(args.refine_config)
 given_ground_truth = data_config.bed is not None  # (benchmarking mode)
 
+
 def call(device, chr_names, uid):
     # runs SV calling on the specified device for the specified list of chromosomes
     # load the pre-trained model on the specified device
@@ -88,7 +89,6 @@ def call(device, chr_names, uid):
                                       collect_data_metrics=True, given_ground_truth=given_ground_truth)
         torch.save(predictions, "%s/predictions.pkl" % predictions_dir)
     return True
-
 
 
 # ------ Image-based discovery ------
@@ -138,6 +138,7 @@ for chr_name in chr_names:
     io.write_bed("%s/candidate_svs.%s.bed" % (config.report_dir, chr_name), chr2calls[chr_name])
 
 # ------ NN-aided breakpoint refinement ------
+post_process_refined = False
 if refine_config is not None and refine_config.pretrained_model is not None:
     def refine(device, chr_names):
         refinet = models.CueModelConfig(refine_config).get_model()
@@ -155,7 +156,8 @@ if refine_config is not None and refine_config.pretrained_model is not None:
             chr_out_bed_file = "%s/refined_svs.%s.bed" % (config.report_dir, chr_name)
             io.write_bed(chr_out_bed_file, chr_calls)
     Parallel(n_jobs=n_procs)(delayed(refine)(chr_name_chunks[i]) for i in range(n_procs))
-else: # ------ Genome-based breakpoint refinement ------
+    post_process_refined = True
+elif not data_config.refine_disable:  # ------ Genome-based breakpoint refinement ------
     def refine(chr_names):
         for chr_name in chr_names:
             chr_calls = io.bed2sv_calls("%s/candidate_svs.%s.bed" % (config.report_dir, chr_name)) 
@@ -163,17 +165,19 @@ else: # ------ Genome-based breakpoint refinement ------
             chr_calls = seq.refinery.refine_svs(chr_calls, data_config, chr_index)
             io.write_bed("%s/refined_svs.%s.bed" % (config.report_dir, chr_name), chr_calls)
     Parallel(n_jobs=n_procs)(delayed(refine)(chr_name_chunks[i]) for i in range(n_procs))
+    post_process_refined = True
     
 
 # output candidate SVs (post-refinement)
-sv_calls_refined = []
-for chr_name in chr_names:
-    sv_calls_refined.extend(io.bed2sv_calls("%s/refined_svs.%s.bed" % (config.report_dir, chr_name)))
-    os.remove("%s/refined_svs.%s.bed" % (config.report_dir, chr_name))
-candidate_out_bed_file = "%s/refined_svs.bed" % config.report_dir
-io.write_bed(candidate_out_bed_file, sv_calls_refined)
+if post_process_refined:
+    sv_calls_refined = []
+    for chr_name in chr_names:
+        sv_calls_refined.extend(io.bed2sv_calls("%s/refined_svs.%s.bed" % (config.report_dir, chr_name)))
+        os.remove("%s/refined_svs.%s.bed" % (config.report_dir, chr_name))
+    candidate_out_bed_file = "%s/refined_svs.bed" % config.report_dir
+    io.write_bed(candidate_out_bed_file, sv_calls_refined)
 
 # ------ IO ------
 # write SV calls to file
 io.bed2vcf(candidate_out_bed_file, "%s/svs.vcf" % config.report_dir, data_config.fai,
-                   min_score=data_config.min_qual_score, min_len=data_config.min_sv_len)
+           min_score=data_config.min_qual_score, min_len=data_config.min_sv_len)
